% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadsubread.R
\name{DEXSeqDataSetFromFeatureCounts}
\alias{DEXSeqDataSetFromFeatureCounts}
\title{Create DEXSeq from feactureCounts}
\usage{
DEXSeqDataSetFromFeatureCounts(countfile, sampleData, design = ~sample + exon
  + condition:exon, flattenedfile = NULL)
}
\arguments{
\item{countfile}{Output from featureCounts}

\item{sampleData}{A data.frame with sample names and conditions (control/treatment)}

\item{design}{How the DEXSeq design matrix should be built. Defaults to `sample + exon + condition:exon`}

\item{flattenedfile}{Annotation file}
}
\description{
This function will take featureCounts output and together with other outputs
will create a DEXSeq design matrix
}
